parse_hint: fix function and add unit test
code review indicated that parse_hint() would loop forever when looking
for "stuff=" in "foostuff=bar" (e.g. stuff found not at start of line or
after a comma)

Add test to illustrate the problem and fix it.

+++ copytool/reporting.c
@@ -70,7 +70,7 @@ int report_new_action(struct hsm_action_node *han)
 
 	const char *hint_needle = state->config.reporting_hint;
 	const char *hint;
-	int data_len;
+	size_t data_len;
 
 	hint = parse_hint(han, hint_needle, &data_len);
 	if (!hint)
@@ -81,7 +81,7 @@ int report_new_action(struct hsm_action_node *han)
 	 * - not alnum, - or _ (in particular, no /, . or nul byte allowed)
 	 */
 	if (data_len > 64) {
-		LOG_INFO("fid " DFID " reporting hint was longer than 64 (%d)",
+		LOG_INFO("fid " DFID " reporting hint was longer than 64 (%zd)",
 			 PFID(&han->info.dfid), data_len);
 		return -EINVAL;
 	}
